Unit test cases added.
