+++ x-pack/plugins/monitoring/server/telemetry_collection/__mocks__/fixtures/logstash_stats_self_monitoring_results.json
@@ -43,9 +43,6 @@
           "_id" : "rC2jSncB1VO1nvgvsKx6",
           "_score" : null,
           "_source" : {
-            "agent" : {
-              "type" : "metricbeat"
-            },
             "logstash_stats" : {
               "logstash" : {
                 "version" : "7.8.0",
